+++ src/set.h
@@ -22,7 +22,7 @@
  *  Thomas Nau, Schlehenweg 15, 88471 Baustetten, Germany
  *  Thomas.Nau@rz.uni-ulm.de
  *
- *  RCS: $Id: set.h,v 1.3 2004-01-15 14:21:29 haceaton Exp $
+ *  RCS: $Id: set.h,v 1.4 2004-01-19 19:16:18 haceaton Exp $
  */
 
 /* prototypes for update routines
@@ -36,7 +36,7 @@
 void		SetCursorStatusLine(void);
 void		SetStatusLine(void);
 void		SetGrid(float, Boolean);
-void		SetZoom(int);
+void		SetZoom(float);
 void		SetLineSize(BDimension);
 void		SetViaSize(BDimension, Boolean);
 void		SetViaDrillingHole(BDimension, Boolean);
